// src/functions/indicateurStrategique/types.ts
export interface typeIndicStrategique {
    id_indicateur_str: number
    niveau_istr: number
    code_indicateur_istr: string 
    code_istr: string 
    intitule_indicateur_istr: string 
    periodicite_iop: string 
    source_istr: string 
    responsable_istr: string 
    description_istr: string 
    structure_istr: string 
    programme_istr: number 
}